moved type definitions to separate file
